Move tooltips/ to common/
