+++ components/cartes-impact/simple-card/simple-card-component.jsx
@@ -18,7 +18,7 @@ import React, { Fragment } from "react";
 
 import { Card } from "../../common";
 import DoublePalmTreeIcon from "../../icons/double-palm-tree";
-import { NeutralTooltip } from "../../tooltips";
+import { NeutralTooltip } from "../../common";
 import { formatNumber } from "../../utils";
 import SimpleCardImpactImpots from "./impact-impots";
