+++ components/common/articles-inputs/parameter/Parameter.tsx
@@ -1,7 +1,7 @@
 import classNames from "classnames";
 import { PureComponent } from "react";
 
-import { PlfTooltip, ReformTooltip } from "../../../tooltips";
+import { PlfTooltip, ReformTooltip } from "../../tooltips";
 import { formatNumber } from "../../../utils";
 import { NumberInput } from "./number-input";
 import styles from "./Parameter.module.scss";
