+++ components/common/index.ts
@@ -2,3 +2,4 @@ export { default as AppHeader } from "./app-header";
 export * from "./articles-inputs";
 export * from "./card"
 export * from "./expandable-panels";
+export * from "./tooltips";
