+++ components/common/tooltips/NeutralTooltip.jsx
@@ -1,4 +1,4 @@
-import { neutralColor } from "../colors";
+import { neutralColor } from "../../colors";
 import { createColoredTooltip } from "./create-colored-tooltip";
 
 export const NeutralTooltip = createColoredTooltip(neutralColor);
