+++ components/common/tooltips/create-colored-tooltip.jsx
